fix: improve test infrastructure

+++ src/tradingbot/api/models/trade.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel, Field
@@ -10,10 +9,10 @@ from .trading import TradeStatus
 class TradeCreate(BaseModel):
     symbol: str
     direction: str
-    quantity: Decimal
-    entry_price: Decimal
-    take_profit: Optional[Decimal] = None
-    stop_loss: Optional[Decimal] = None
+    quantity: float
+    entry_price: float
+    take_profit: Optional[float] = None
+    stop_loss: Optional[float] = None
     leverage: Optional[float] = Field(default=1.0)
     metadata: dict = Field(default_factory=dict)
 
@@ -21,11 +20,11 @@ class Trade(BaseModel):
     id: PyObjectId = Field(default_factory=PyObjectId, alias="_id")
     symbol: str
     direction: str
-    quantity: Decimal
-    entry_price: Decimal
-    exit_price: Optional[Decimal] = None
-    take_profit: Optional[Decimal] = None
-    stop_loss: Optional[Decimal] = None
+    quantity: float
+    entry_price: float
+    exit_price: Optional[float] = None
+    take_profit: Optional[float] = None
+    stop_loss: Optional[float] = None
     leverage: float = Field(default=1.0)
     status: TradeStatus = Field(default=TradeStatus.PENDING)
     metadata: dict = Field(default_factory=dict)
